requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
drf-yasg
Pillow
django_heroku
gunicorn
psycopg2
whitenoise